requirements.txt
streamlit
pandas
python-dotenv
matplotlib
numpy